move shared memory helpers to `tensorrtserver.shared_memory`
